requirements.txt
argparse
dill
fire
matplotlib
numba
psutil
requests
scikit-image
scipy
slidingwindow
tqdm
git+https://github.com/ppwwyyxx/tensorpack.git
pyqt5-tools